EWL-6567: Update select pattern and add jump menu JS.
The pattern seems to be used as a stand in for multiple different select boxes. There should probably be multiple patterns, but I'm not sure right now what they are and what would be affected.

I've changed `js-dropdown-select` to `ama__jump_menu` in order to match the existing pattern with the use of the `data-url` attributes. This is the jump menu pattern. The `js-dropdown-select` is also used for exposed filters and other select boxes. The comment is intended to help a bit there.

The jump menu JS is largely taken from the previous rivendale theme code, but the selector is changed to `ama__jump_menu` and the event `selectmenuchange` was used instead of `change` since I had more success with that.